Extract reactive bits into `ui.ReactiveInterface`
